Fix for blank icons on Galaxy Note 2.
The icon remained completely blank on the Galaxy Note 2 (GT-N7100) running
Android 4.1.2. It's a weird one, because everything seems to be in order (e.g.
the path bounds come out correctly if you log them), but the path doesn't
actually get drawn and the canvas remains transparent.

While investigating the issue, I stumbled upon adding a bogus path point as a
fix for the problem. I have no idea why this is needed or why it works, but
I'll take what I can get.

If you scale the icon really huge, you may see a colored pixel in the top left.